Added subfolder Builtin with relative builtins. TO-DO: Unset. Modified makefile and header file with relative added functions
