+++ srcs/builtins/unset.c
@@ -1,0 +1,9 @@
+#include "libft.h"
+#include "minishell.h"
+#include <stdlib.h>
+#include <unistd.h>
+
+// void	unset_variable(t_data *data)
+// {
+// 	return ;
+// }
